arm64: crypto: Improve target optimization flags
The target flags which are needed to enable
extensions for cpu intructions were overriding
our global setting of "mcpu=cortex-a57" and targeting
a generic armv8-s cpus instead.

This patch adjusts these flags to target
"cortex-a57", which is the closest cpu target
to kryo, while keeping the extensions for
cpu instructions.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>